benchmark/meilisearch/requirements.txt
annotated-types==0.7.0
beautifulsoup4==4.14.3
camel-converter==5.1.0
certifi==2026.2.25
charset-normalizer==3.4.6
idna==3.11
meilisearch==0.40.0
pydantic==2.12.5
pydantic-core==2.41.5
pymysql==1.1.2
python-dotenv==1.2.2
requests==2.32.5
soupsieve==2.8.3
tabulate==0.10.0
tqdm==4.67.3
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
